Avoid global lock of BATsample by using private seed.

+++ gdk/gdk.h
@@ -2199,7 +2199,7 @@ gdk_export gdk_return BATfirstn(BAT **topn, BAT **gids, BAT *b, BAT *cands, BAT 
  *
  */
 gdk_export BAT *BATsample(BAT *b, BUN n);
-gdk_export BAT *BATsample_with_seed(BAT *b, BUN n, unsigned seed);
+gdk_export BAT *BATsample_with_seed(BAT *b, BUN n, uint64_t seed);
 
 /*
  *
